Rename DB ID-generation file and clear session-populated tables before updating
